Move sdk_extras module
